Feature/server 17

+++ src/main/java/privacy/controllers/FriendshipResponseController.java
@@ -3,6 +3,7 @@ package privacy.controllers;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.boot.configurationprocessor.json.JSONObject;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -14,32 +15,17 @@ import privacy.dao.OwnerRepository;
 import privacy.general.payload.response.FriendshipResponse;
 import privacy.general.payload.response.FriendshipResponseItem;
 import privacy.models.Friendship;
+import privacy.models.Owner;
 import privacy.models.new_friend.FriendshipRequestAccepted;
 import privacy.models.new_friend.FriendshipRequestCreated;
 import privacy.service.security.jwt.AuthEntryPointJwt;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import static privacy.models.new_friend.EStatus.ACCEPT;
 
-/** Friendship requests and answers:
-
- Send a request as user 1 (for example), on localhost:8080/api/send_new_fr_request:
- {
- "receiverId": 4,
- "publicKey": "publicKeyFrom6to8"
- }
- As user 4:
- Check your requests on localhost:8080/api/received_fr_requests; //the status of the request is set to PENDING in the requests table
- Answer a request on localhost:8080/api/answer_new_fr_request:
- {
- "frInitiatorId":1,
- "symmetricKey":"symmetricKey4For1",
- "status":"ACCEPT" //or REJECT, in which case - no symmetric key will be saved
- }
- The sent request gets deleted from the firs table - fr_request_created - and the answer is saved in the second table
- - of fr_request_accepted, containing the status - ACCEPT or REJECT */
 @CrossOrigin(origins = "*", maxAge = 3600)
 @RestController
 @RequiredArgsConstructor
@@ -70,35 +56,44 @@ public class FriendshipResponseController{
         Long currentUserId = ownerRepository.findOwnerByUsername(currentUser).get().getOwnerId();
 
         frResponse.setRequestAccepter(currentUserId);
+        Optional<Owner> initiatorUser = ownerRepository.findById(frResponse.getFrInitiatorId());
+        if (initiatorUser.isPresent() && friendshipRequestsRepository.existsFriendshipRequestCreatedBySenderIdAndAndReceiverId(frResponse.getFrInitiatorId(), frResponse.getRequestAccepter())) {
+            frResponse.setInitiatorUsername(initiatorUser.get().getUsername());
 
-        if (frResponse.getStatus().name() == "ACCEPT") {
+            if (frResponse.getStatus().name() == "ACCEPT") {
 
-            FriendshipRequestAccepted friendshipRequestResponse = new FriendshipRequestAccepted(
-                    frResponse.getResponseToRequestId(),
-                    frResponse.getFrInitiatorId(),
-                    frResponse.getRequestAccepter(),
-                    frResponse.getSymmetricKey(),
-                    frResponse.getStatus()); //get fr id
+                FriendshipRequestAccepted friendshipRequestResponse = new FriendshipRequestAccepted(
+                        frResponse.getResponseToRequestId(),
+                        frResponse.getFrInitiatorId(),
+                        frResponse.getInitiatorUsername(),
+                        frResponse.getRequestAccepter(),
+                        frResponse.getSymmetricKey(),
+                        frResponse.getStatus());
 
-            friendshipResponsesRepository.save(friendshipRequestResponse);
-            FriendshipRequestCreated answered = friendshipRequestsRepository.findFriendshipRequestCreatedBySenderIdAndReceiverId(frResponse.getFrInitiatorId(), frResponse.getRequestAccepter());
-            answered.setStatus(ACCEPT);
-            friendshipRequestsRepository.delete(answered);
-            Friendship newFriendship = new Friendship(friendshipRequestResponse.getFrInitiatorId(), friendshipRequestResponse.getRequestAccepter());
-            friendshipRepository.save(newFriendship);
+                friendshipResponsesRepository.save(friendshipRequestResponse);
+                FriendshipRequestCreated answered = friendshipRequestsRepository.findFriendshipRequestCreatedBySenderIdAndReceiverId(frResponse.getFrInitiatorId(), frResponse.getRequestAccepter());
+                answered.setStatus(ACCEPT);
+                friendshipRequestsRepository.delete(answered);
+                Friendship newFriendship = new Friendship(friendshipRequestResponse.getFrInitiatorId(), friendshipRequestResponse.getRequestAccepter());
+                friendshipRepository.save(newFriendship);
 
-            return ResponseEntity.ok(newFriendship);
-        }else{
-            FriendshipRequestAccepted friendshipRequestResponse = new FriendshipRequestAccepted(
-                    frResponse.getResponseToRequestId(),
-                    frResponse.getRequestAccepter(),
-                    frResponse.getFrInitiatorId(),
-                    frResponse.getSymmetricKey().replaceAll(frResponse.getSymmetricKey(), ""),
-                    frResponse.getStatus());
-            friendshipResponsesRepository.save(friendshipRequestResponse);
-            FriendshipRequestCreated answered = friendshipRequestsRepository.findFriendshipRequestCreatedBySenderIdAndReceiverId(frResponse.getFrInitiatorId(), frResponse.getRequestAccepter());
-            friendshipRequestsRepository.delete(answered);
-            return ResponseEntity.ok(friendshipRequestResponse);
+                return ResponseEntity.ok(newFriendship);
+            } else {
+                FriendshipRequestAccepted friendshipRequestResponse = new FriendshipRequestAccepted(
+                        frResponse.getResponseToRequestId(),
+                        frResponse.getFrInitiatorId(),
+                        frResponse.getInitiatorUsername(),
+                        frResponse.getRequestAccepter(),
+                        frResponse.getSymmetricKey().replaceAll(frResponse.getSymmetricKey(), ""),
+                        frResponse.getStatus());
+                friendshipResponsesRepository.save(friendshipRequestResponse);
+                FriendshipRequestCreated answered = friendshipRequestsRepository.findFriendshipRequestCreatedBySenderIdAndReceiverId(frResponse.getFrInitiatorId(), frResponse.getRequestAccepter());
+                friendshipRequestsRepository.delete(answered);
+                return ResponseEntity.ok(friendshipRequestResponse);
+            }
+        }
+        else {
+            return ResponseEntity.status(404).body("Error: friendship initiator with this id found");
         }
     }
 
@@ -135,8 +130,10 @@ public class FriendshipResponseController{
 
             for (FriendshipRequestAccepted item:
                     responsesList) {
-                Friendship friendship = friendshipRepository.findFriendshipByUserOneIdAndUserTwoId(initiatorId, item.getRequestAccepter());
-                if (friendship != null){
+                Optional<Friendship> friendshipTry = friendshipRepository.findFriendshipByUserOneIdAndUserTwoId(initiatorId, item.getRequestAccepter());
+                if (friendshipTry.isPresent()){
+                    Friendship friendship = friendshipTry.get();
+                    item.setAccepterUsername(ownerRepository.findById(item.getRequestAccepter()).get().getUsername());
                     FriendshipResponseItem new_response = new FriendshipResponseItem(friendship, item);
                     responsesListItems.add(new_response);
                 }
@@ -149,7 +146,7 @@ public class FriendshipResponseController{
             return new ResponseEntity<>(friendshipResponse, HttpStatus.OK);
         } catch (Exception e) {
             logger.error("Couldn't obtain responses to friendship requests for user "+ currentUser);
-            return ResponseEntity.status(204).body("An unexpected error occurred");
+            return ( ResponseEntity.status(204).body("An unexpected error occurred"));
         }
     }
 }
